feat: 不要なdomainを削除
